✅Update goerli deploy scripts
